Adding 2023 Day 06 in C# (#15)
* Adding 2023 Day 06 in C#

* Completing Part One

* Completing Part Two

+++ 2023/csharp/Advent23.Days/DayTemplate/Day00Test.cs
@@ -5,8 +5,8 @@ namespace Advent23.Days
 {
     public class Day00Test : TestEngine<Day00Solver, object[], long>
     {
-        private const int EXPECTED_SOLUTION_PART_1 = 123;
-        private const int EXPECTED_SOLUTION_PART_2 = 123;
+        private const long EXPECTED_SOLUTION_PART_1 = 123;
+        private const long EXPECTED_SOLUTION_PART_2 = 123;
 
         public Day00Test()
         {
